Added Blockly.TopComponent to sit above both the html & svg trees.
Had a hangout with Neil last night - ultimately Blockly might be an object that can have multiple instances that can be injected wherever the caller wants.

For the moment as a transition, I'm providing TopComponent to manage both the html pieces and the svg pieces (through its SvgComponent child).

Once all the ui objects are Components in the tree somewhere under Blockly.TopComponent, we can focus on migrating "Blockly" from a namespace to a constructor/object that can be instantiated.

TBR=neil.fraser@gmail.com
BUG=103

Review URL: https://codereview.appspot.com/7307089

git-svn-id: https://blockly.googlecode.com/svn/trunk@633 c400ca83-b69d-9dd7-9705-49c6b8615e23